refactor: move onClose handling to shared utility hook
